isoler les types dans fichier
